[master temperature_humidity] Fixes the datatype isssue.

+++ api/v1/models/weather.go
@@ -3,6 +3,6 @@ package models
 type WeatherMessage struct {
 	Success     string
 	Message     string
-	Temperature float64
-	Humidity    float64
+	Temperature float32
+	Humidity    float32
 }
